+++ src/test/java/picoded/dstack/redisson/Redisson_KeyValueMap_test.java
@@ -1,4 +1,4 @@
-package picoded.dstack.redis;
+package picoded.dstack.redisson;
 
 import picoded.core.struct.*;
 import picoded.dstack.*;
@@ -8,11 +8,11 @@ import java.util.concurrent.ThreadLocalRandom;
 
 /**
  * ## Purpose
- * This class is meant to test the Redis_KeyValueMap implementation,
+ * This class is meant to test the Redisson_KeyValueMap implementation,
  * and ensure that it passes all the test layed out in StructSimple_KeyValueMap_test
  * 
  */
-public class Redis_KeyValueMap_test extends StructSimple_KeyValueMap_test {
+public class Redisson_KeyValueMap_test extends StructSimple_KeyValueMap_test {
 	// Redis stack instance
 	protected static volatile RedisStack instance = null;
 	
@@ -20,7 +20,7 @@ public class Redis_KeyValueMap_test extends StructSimple_KeyValueMap_test {
 	public KeyValueMap implementationConstructor() {
 		
 		// Initialize server
-		synchronized (Redis_KeyValueMap_test.class) {
+		synchronized (Redisson_KeyValueMap_test.class) {
 			if (instance == null) {
 				// The default config uses "172.17.0.1" (default docker bridge address) 
 				// and port 6379 (default redis port)
@@ -33,7 +33,7 @@ public class Redis_KeyValueMap_test extends StructSimple_KeyValueMap_test {
 				redisConfig.put("name", DStackTestConfig.randomTablePrefix());
 				
 				GenericConvertMap<String, Object> stackConfig = new GenericConvertHashMap<>();
-				stackConfig.put("name", "Redis_KeyValueMap_test");
+				stackConfig.put("name", "Redisson_KeyValueMap_test");
 				stackConfig.put("redis", redisConfig);
 				
 				instance = new RedisStack(stackConfig);
